# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.3
info:
  title: API de Catálogos
  description: API REST para gestionar catálogos usando arquitectura hexagonal con Quarkus
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
paths:
  /api/catalogos:
    post:
      summary: Crear un nuevo catálogo
      description: Crea un nuevo catálogo con los detalles proporcionados.
      operationId: crearCatalogo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Catalogo'
      responses:
        '201':
          description: Catálogo creado exitosamente
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: URI del catálogo recién creado
        '400':
          description: Solicitud inválida
    get:
      summary: Buscar catálogo por descripción
      description: Busca un catálogo por su descripción exacta.
      operationId: buscarCatalogoPorDescripcion
      parameters:
        - name: descripcion
          in: query
          required: true
          schema:
            type: string
          description: Descripción del catálogo a buscar
      responses:
        '200':
          description: Catálogo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalogo'
        '400':
          description: Descripción no proporcionada o vacía
        '404':
          description: Catálogo no encontrado
  /api/catalogos/{id}:
    get:
      summary: Obtener catálogo por ID
      description: Obtiene un catálogo específico por su ID.
      operationId: obtenerCatalogo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID del catálogo a obtener
      responses:
        '200':
          description: Catálogo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Catalogo'
        '400':
          description: ID inválido
        '404':
          description: Catálogo no encontrado
    put:
      summary: Modificar un catálogo existente
      description: Modifica un catálogo existente identificado por su ID.
      operationId: modificarCatalogo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
          description: ID del catálogo a modificar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Catalogo'
      responses:
        '204':
          description: Catálogo modificado exitosamente
        '400':
          description: Solicitud inválida
        '404':
          description: Catálogo no encontrado
  /api/catalogos/paginated/list:
    get:
      summary: Obtener catálogos paginados
      description: Obtiene una lista paginada de todos los catálogos. Los parámetros de paginación son opcionales, con valores por defecto de página 0 y 10 items por página.
      operationId: listarCatalogosPaginado
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
          description: Número de página (comienza en 0). Por defecto es 0.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
            maximum: 100
          description: Cantidad de items por página. Por defecto es 10. Máximo 100 items.
      responses:
        '200':
          description: Lista paginada de catálogos obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponseCatalogo'
        '400':
          description: Parámetros de paginación inválidos
components:
  schemas:
    Catalogo:
      type: object
      description: Representa un catálogo en el sistema
      properties:
        codCat:
          type: integer
          format: int32
          readOnly: true
          description: Código único del catálogo (generado automáticamente)
        catDesc:
          type: string
          description: Descripción del catálogo
          example: Menu Principal
        codUsr:
          type: string
          description: Código de usuario que gestiona el catálogo
          example: rsanchez
        fecha:
          type: string
          format: date-time
          readOnly: true
          description: Fecha de creación o última modificación (generada automáticamente)
          example: 2025-11-12T19:48:00Z
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemCatalogo'
          description: Lista de items que contiene el catálogo
      required:
        - catDesc
        - codUsr
    ItemCatalogo:
      type: object
      description: Representa un item dentro de un catálogo
      properties:
        codCor:
          type: integer
          format: int32
          description: Código único del ítem
          example: 1
        cirDesc:
          type: string
          description: Descripción del ítem
          example: Opción 1 del menú
      required:
        - codCor
        - cirDesc
    PaginatedResponseCatalogo:
      type: object
      description: Respuesta paginada con lista de catálogos
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Catalogo'
          description: Lista de catálogos en la página actual
        page:
          type: integer
          format: int32
          description: Número de página actual (comienza en 0)
          example: 0
        size:
          type: integer
          format: int32
          description: Cantidad de items por página
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Total de catálogos disponibles en toda la base de datos
          example: 25
        totalPages:
          type: integer
          format: int32
          description: Total de páginas disponibles
          example: 3
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages